requirements.txt
anyio==4.12.1
APScheduler==3.11.2
astroid==4.0.4
beautifulsoup4==4.14.3
black==26.3.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
coverage==7.13.4
dill==0.4.1
feedparser==6.0.12
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==8.0.1
librt==0.8.1
mccabe==0.7.0
mypy==1.19.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pylint==4.0.5
pytest==9.0.
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dotenv==1.2.2
python-telegram-bot==21.10
pytokens==0.4.1
requests==2.32.5
sgmllib3k==1.0.0
soupsieve==2.8.3
tomlkit==0.14.0
types-requests==2.32.4.20260107
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
